r/image: refactoring the tests to use a template which should make this easier to refactor
