+++ azurerm/internal/services/compute/image_resource_test.go
@@ -1,12 +1,8 @@
 package compute_test
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"log"
-	"net/http"
-	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
@@ -16,7 +12,6 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance/check"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
-	"golang.org/x/crypto/ssh"
 )
 
 type ImageResource struct {
@@ -35,14 +30,14 @@ func TestAccImage_standaloneImage(t *testing.T) {
 	data.ResourceTest(t, r, []resource.TestStep{
 		{
 			// need to create a vm and then reference it in the image creation
-			Config: r.standaloneImage_setup(data, userName, password, hostName, "LRS"),
+			Config: r.setupUnmanagedDisks(data, "LRS"),
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMExists("azurerm_virtual_machine.testsource", true),
 				testGeneralizeVMImage(resourceGroup, "testsource", userName, password, hostName, sshPort, data.Locations.Primary),
 			),
 		},
 		{
-			Config: r.standaloneImage_provision(data, userName, password, hostName, "LRS", ""),
+			Config: r.standaloneImageProvision(data, "LRS", ""),
 			Check: resource.ComposeTestCheckFunc(
 				check.That("azurerm_image.test").ExistsInAzure(r),
 			),
@@ -66,14 +61,14 @@ func TestAccImage_standaloneImage_hyperVGeneration_V2(t *testing.T) {
 	data.ResourceTest(t, r, []resource.TestStep{
 		{
 			// need to create a vm and then reference it in the image creation
-			Config: r.standaloneImage_setup(data, userName, password, hostName, "LRS"),
+			Config: r.setupUnmanagedDisks(data, "LRS"),
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMExists("azurerm_virtual_machine.testsource", true),
 				testGeneralizeVMImage(resourceGroup, "testsource", userName, password, hostName, sshPort, data.Locations.Primary),
 			),
 		},
 		{
-			Config: r.standaloneImage_provision(data, userName, password, hostName, "LRS", "V2"),
+			Config: r.standaloneImageProvision(data, "LRS", "V2"),
 			Check: resource.ComposeTestCheckFunc(
 				check.That("azurerm_image.test").ExistsInAzure(r),
 			),
@@ -97,7 +92,7 @@ func TestAccImage_standaloneImageZoneRedundant(t *testing.T) {
 	data.ResourceTest(t, r, []resource.TestStep{
 		{
 			// need to create a vm and then reference it in the image creation
-			Config:  r.standaloneImage_setup(data, userName, password, hostName, "ZRS"),
+			Config:  r.setupUnmanagedDisks(data, "ZRS"),
 			Destroy: false,
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMExists("azurerm_virtual_machine.testsource", true),
@@ -105,7 +100,7 @@ func TestAccImage_standaloneImageZoneRedundant(t *testing.T) {
 			),
 		},
 		{
-			Config: r.standaloneImage_provision(data, userName, password, hostName, "ZRS", ""),
+			Config: r.standaloneImageProvision(data, "ZRS", ""),
 			Check: resource.ComposeTestCheckFunc(
 				check.That("azurerm_image.test").ExistsInAzure(r),
 			),
@@ -127,26 +122,26 @@ func TestAccImage_requiresImport(t *testing.T) {
 	data.ResourceTest(t, r, []resource.TestStep{
 		{
 			// need to create a vm and then reference it in the image creation
-			Config: r.standaloneImage_setup(data, userName, password, hostName, "LRS"),
+			Config: r.setupUnmanagedDisks(data, "LRS"),
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMExists("azurerm_virtual_machine.testsource", true),
 				testGeneralizeVMImage(resourceGroup, "testsource", userName, password, hostName, sshPort, data.Locations.Primary),
 			),
 		},
 		{
-			Config: r.standaloneImage_provision(data, userName, password, hostName, "LRS", ""),
+			Config: r.standaloneImageProvision(data, "LRS", ""),
 			Check: resource.ComposeTestCheckFunc(
 				check.That("azurerm_image.test").ExistsInAzure(r),
 			),
 		},
 		{
-			Config:      r.standaloneImage_requiresImport(data, userName, password, hostName),
+			Config:      r.standaloneImageRequiresImport(data),
 			ExpectError: acceptance.RequiresImportError("azurerm_image"),
 		},
 	})
 }
 
-func TestAccImage_customImageVMFromVHD(t *testing.T) {
+func TestAccImage_customImageFromVMWithUnmanagedDisks(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_virtual_machine", "testsource")
 	r := ImageResource{}
 
@@ -159,14 +154,14 @@ func TestAccImage_customImageVMFromVHD(t *testing.T) {
 	data.ResourceTest(t, r, []resource.TestStep{
 		{
 			// need to create a vm and then reference it in the image creation
-			Config: r.customImage_fromVHD_setup(data, userName, password, hostName),
+			Config: r.setupUnmanagedDisks(data, "LRS"),
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMExists("azurerm_virtual_machine.testsource", true),
 				testGeneralizeVMImage(resourceGroup, "testsource", userName, password, hostName, sshPort, data.Locations.Primary),
 			),
 		},
 		{
-			Config: r.customImage_fromVHD_provision(data, userName, password, hostName),
+			Config: r.customImageFromVMWithUnmanagedDisksProvision(data),
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMExists("azurerm_virtual_machine.testdestination", true),
 			),
@@ -174,7 +169,7 @@ func TestAccImage_customImageVMFromVHD(t *testing.T) {
 	})
 }
 
-func TestAccImage_customImageVMFromVM(t *testing.T) {
+func TestAccImage_customImageFromVMWithManagedDisks(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_virtual_machine", "testsource")
 	r := ImageResource{}
 
@@ -187,7 +182,7 @@ func TestAccImage_customImageVMFromVM(t *testing.T) {
 	data.ResourceTest(t, r, []resource.TestStep{
 		{
 			// need to create a vm and then reference it in the image creation
-			Config:  r.customImage_fromVM_sourceVM(data, userName, password, hostName),
+			Config:  r.setupManagedDisks(data),
 			Destroy: false,
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMExists("azurerm_virtual_machine.testsource", true),
@@ -195,7 +190,7 @@ func TestAccImage_customImageVMFromVM(t *testing.T) {
 			),
 		},
 		{
-			Config: r.customImage_fromVM_destinationVM(data, userName, password, hostName),
+			Config: r.customImageFromManagedDiskVMProvision(data),
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMExists("azurerm_virtual_machine.testdestination", true),
 			),
@@ -203,7 +198,7 @@ func TestAccImage_customImageVMFromVM(t *testing.T) {
 	})
 }
 
-func TestAccImageVMSS_customImageVMSSFromVHD(t *testing.T) {
+func TestAccImage_customImageFromVMSSWithUnmanagedDisks(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azurerm_virtual_machine", "testsource")
 	r := ImageResource{}
 
@@ -216,7 +211,7 @@ func TestAccImageVMSS_customImageVMSSFromVHD(t *testing.T) {
 	data.ResourceTest(t, r, []resource.TestStep{
 		{
 			// need to create a vm and then reference it in the image creation
-			Config:  r.vmss_customImage_fromVHD_setup(data, userName, password, hostName),
+			Config:  r.setupUnmanagedDisks(data, "LRS"),
 			Destroy: false,
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMExists("azurerm_virtual_machine.testsource", true),
@@ -224,7 +219,7 @@ func TestAccImageVMSS_customImageVMSSFromVHD(t *testing.T) {
 			),
 		},
 		{
-			Config: r.vmss_customImage_fromVHD_provision(data, userName, password, hostName),
+			Config: r.customImageFromVMSSWithUnmanagedDisksProvision(data),
 			Check: resource.ComposeTestCheckFunc(
 				testCheckAzureVMSSExists("azurerm_virtual_machine_scale_set.testdestination", true),
 			),
@@ -232,110 +227,7 @@ func TestAccImageVMSS_customImageVMSSFromVHD(t *testing.T) {
 	})
 }
 
-// nolint unparam
-func testGeneralizeVMImage(resourceGroup string, vmName string, userName string, password string, hostName string, port string, location string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		armClient := acceptance.AzureProvider.Meta().(*clients.Client)
-		vmClient := armClient.Compute.VMClient
-		ctx := armClient.StopContext
-
-		normalizedLocation := azure.NormalizeLocation(location)
-		suffix := armClient.Account.Environment.ResourceManagerVMDNSSuffix
-		dnsName := fmt.Sprintf("%s.%s.%s", hostName, normalizedLocation, suffix)
-
-		if err := deprovisionVM(userName, password, dnsName, port); err != nil {
-			return fmt.Errorf("Bad: Deprovisioning error %+v", err)
-		}
-
-		future, err := vmClient.Deallocate(ctx, resourceGroup, vmName)
-		if err != nil {
-			return fmt.Errorf("Bad: Deallocating error %+v", err)
-		}
-
-		if err = future.WaitForCompletionRef(ctx, vmClient.Client); err != nil {
-			return fmt.Errorf("Bad: Deallocating error %+v", err)
-		}
-
-		if _, err = vmClient.Generalize(ctx, resourceGroup, vmName); err != nil {
-			return fmt.Errorf("Bad: Generalizing error %+v", err)
-		}
-
-		return nil
-	}
-}
-
-func deprovisionVM(userName string, password string, hostName string, port string) error {
-	// SSH into the machine and execute a waagent deprovisioning command
-	var b bytes.Buffer
-	cmd := "sudo waagent -verbose -deprovision+user -force"
-
-	config := &ssh.ClientConfig{
-		User: userName,
-		Auth: []ssh.AuthMethod{
-			ssh.Password(password),
-		},
-		HostKeyCallback: ssh.InsecureIgnoreHostKey(),
-	}
-	log.Printf("[INFO] Connecting to %s:%v remote server...", hostName, port)
-
-	hostAddress := strings.Join([]string{hostName, port}, ":")
-	client, err := ssh.Dial("tcp", hostAddress, config)
-	if err != nil {
-		return fmt.Errorf("Bad: deprovisioning error %+v", err)
-	}
-
-	session, err := client.NewSession()
-	if err != nil {
-		return fmt.Errorf("Bad: deprovisioning error, failure creating session %+v", err)
-	}
-	defer session.Close()
-
-	session.Stdout = &b
-	if err := session.Run(cmd); err != nil {
-		return fmt.Errorf("Bad: deprovisioning error, failure running command %+v", err)
-	}
-
-	return nil
-}
-
-// nolint unparam
-func testCheckAzureVMExists(sourceVM string, shouldExist bool) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		client := acceptance.AzureProvider.Meta().(*clients.Client).Compute.VMClient
-		ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
-
-		log.Printf("[INFO] testing MANAGED IMAGE VM EXISTS - BEGIN.")
-
-		vmRs, vmOk := s.RootModule().Resources[sourceVM]
-		if !vmOk {
-			return fmt.Errorf("VM Not found: %s", sourceVM)
-		}
-		vmName := vmRs.Primary.Attributes["name"]
-
-		resourceGroup, hasResourceGroup := vmRs.Primary.Attributes["resource_group_name"]
-		if !hasResourceGroup {
-			return fmt.Errorf("Bad: no resource group found in state for VM: %s", vmName)
-		}
-
-		resp, err := client.Get(ctx, resourceGroup, vmName, "")
-		if err != nil {
-			return fmt.Errorf("Bad: Get on client: %+v", err)
-		}
-
-		if resp.StatusCode == http.StatusNotFound && shouldExist {
-			return fmt.Errorf("Bad: VM %q (resource group %q) does not exist", vmName, resourceGroup)
-		}
-		if resp.StatusCode != http.StatusNotFound && !shouldExist {
-			return fmt.Errorf("Bad: VM %q (resource group %q) still exists", vmName, resourceGroup)
-		}
-
-		log.Printf("[INFO] testing MANAGED IMAGE VM EXISTS - END.")
-
-		return nil
-	}
-}
-
-func (t ImageResource) Exists(ctx context.Context, clients *clients.Client, state *terraform.InstanceState) (*bool, error) {
+func (ImageResource) Exists(ctx context.Context, clients *clients.Client, state *terraform.InstanceState) (*bool, error) {
 	id, err := azure.ParseAzureResourceID(state.ID)
 	if err != nil {
 		return nil, err
@@ -351,77 +243,77 @@ func (t ImageResource) Exists(ctx context.Context, clients *clients.Client, stat
 	return utils.Bool(resp.ID != nil), nil
 }
 
-func testCheckAzureVMSSExists(sourceVMSS string, shouldExist bool) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		vmssClient := acceptance.AzureProvider.Meta().(*clients.Client).Compute.VMScaleSetClient
-		ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
-
-		log.Printf("[INFO] testing MANAGED IMAGE VMSS EXISTS - BEGIN.")
-
-		vmRs, vmOk := s.RootModule().Resources[sourceVMSS]
-		if !vmOk {
-			return fmt.Errorf("VMSS Not found: %s", sourceVMSS)
-		}
-		vmssName := vmRs.Primary.Attributes["name"]
-
-		resourceGroup, hasResourceGroup := vmRs.Primary.Attributes["resource_group_name"]
-		if !hasResourceGroup {
-			return fmt.Errorf("Bad: no resource group found in state for VMSS: %s", vmssName)
-		}
-
-		resp, err := vmssClient.Get(ctx, resourceGroup, vmssName)
-		if err != nil {
-			return fmt.Errorf("Bad: Get on vmssClient: %+v", err)
-		}
-
-		if resp.StatusCode == http.StatusNotFound && shouldExist {
-			return fmt.Errorf("Bad: VMSS %q (resource group %q) does not exist", vmssName, resourceGroup)
-		}
-		if resp.StatusCode != http.StatusNotFound && !shouldExist {
-			return fmt.Errorf("Bad: VMSS %q (resource group %q) still exists", vmssName, resourceGroup)
-		}
-
-		log.Printf("[INFO] testing MANAGED IMAGE VMSS EXISTS - END.")
-
-		return nil
-	}
-}
-
-func (ImageResource) standaloneImage_setup(data acceptance.TestData, userName string, password string, hostName string, storageType string) string {
+func (r ImageResource) setupManagedDisks(data acceptance.TestData) string {
+	template := r.template(data)
 	return fmt.Sprintf(`
 provider "azurerm" {
   features {}
 }
 
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-%d"
-  location = "%s"
-}
-
-resource "azurerm_virtual_network" "test" {
-  name                = "acctvn-%d"
-  address_space       = ["10.0.0.0/16"]
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-}
-
-resource "azurerm_subnet" "test" {
-  name                 = "acctsub-%d"
-  resource_group_name  = azurerm_resource_group.test.name
-  virtual_network_name = azurerm_virtual_network.test.name
-  address_prefix       = "10.0.2.0/24"
-}
-
-resource "azurerm_public_ip" "test" {
-  name                = "acctpip-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-  allocation_method   = "Dynamic"
-  domain_name_label   = "%s"
-}
+%s
 
 resource "azurerm_network_interface" "testsource" {
-  name                = "acctnicsource-%d"
+  name                = "acctnicsource-${local.number}"
+  location            = azurerm_resource_group.test.location
+  resource_group_name = azurerm_resource_group.test.name
+
+  ip_configuration {
+    name                          = "testconfigurationsource"
+    subnet_id                     = azurerm_subnet.test.id
+    private_ip_address_allocation = "Dynamic"
+    public_ip_address_id          = azurerm_public_ip.test.id
+  }
+}
+
+resource "azurerm_virtual_machine" "testsource" {
+  name                  = "testsource"
+  location              = azurerm_resource_group.test.location
+  resource_group_name   = azurerm_resource_group.test.name
+  network_interface_ids = [azurerm_network_interface.testsource.id]
+  vm_size               = "Standard_D1_v2"
+
+  storage_image_reference {
+    publisher = "Canonical"
+    offer     = "UbuntuServer"
+    sku       = "18.04-LTS"
+    version   = "latest"
+  }
+
+  storage_os_disk {
+    name          = "myosdisk1"
+    caching       = "ReadWrite"
+    create_option = "FromImage"
+  }
+
+  os_profile {
+    computer_name  = "mdimagetestsource"
+    admin_username = local.admin_username
+    admin_password = local.admin_password
+  }
+
+  os_profile_linux_config {
+    disable_password_authentication = false
+  }
+
+  tags = {
+    environment = "Dev"
+    cost-center = "Ops"
+  }
+}
+`, template)
+}
+
+func (r ImageResource) setupUnmanagedDisks(data acceptance.TestData, storageType string) string {
+	template := r.template(data)
+	return fmt.Sprintf(`
+provider "azurerm" {
+  features {}
+}
+
+%s
+
+resource "azurerm_network_interface" "testsource" {
+  name                = "acctnicsource-${local.number}"
   location            = azurerm_resource_group.test.location
   resource_group_name = azurerm_resource_group.test.name
 
@@ -434,16 +326,12 @@ resource "azurerm_network_interface" "testsource" {
 }
 
 resource "azurerm_storage_account" "test" {
-  name                     = "accsa%d"
+  name                     = "accsa${var.random_string}"
   resource_group_name      = azurerm_resource_group.test.name
   location                 = azurerm_resource_group.test.location
   account_tier             = "Standard"
   account_replication_type = "%s"
   allow_blob_public_access = true
-
-  tags = {
-    environment = "Dev"
-  }
 }
 
 resource "azurerm_storage_container" "test" {
@@ -452,6 +340,7 @@ resource "azurerm_storage_container" "test" {
   container_access_type = "blob"
 }
 
+# NOTE: using the legacy vm resource since this test requires an unmanaged disk
 resource "azurerm_virtual_machine" "testsource" {
   name                  = "testsource"
   location              = azurerm_resource_group.test.location
@@ -476,28 +365,24 @@ resource "azurerm_virtual_machine" "testsource" {
 
   os_profile {
     computer_name  = "mdimagetestsource"
-    admin_username = "%s"
-    admin_password = "%s"
+    admin_username = local.admin_username
+    admin_password = local.admin_password
   }
 
   os_profile_linux_config {
     disable_password_authentication = false
   }
-
-  tags = {
-    environment = "Dev"
-    cost-center = "Ops"
-  }
 }
-`, data.RandomInteger, data.Locations.Primary, data.RandomInteger, data.RandomInteger, data.RandomInteger, hostName, data.RandomInteger, data.RandomInteger, storageType, userName, password)
+`, template, storageType)
 }
 
-func (r ImageResource) standaloneImage_provision(data acceptance.TestData, userName string, password string, hostName string, storageType string, hyperVGen string) string {
+func (r ImageResource) standaloneImageProvision(data acceptance.TestData, storageType string, hyperVGen string) string {
 	hyperVGenAtt := ""
 	if hyperVGen != "" {
 		hyperVGenAtt = fmt.Sprintf(`hyper_v_generation = "%s"`, hyperVGen)
 	}
 
+	template := r.setupUnmanagedDisks(data, storageType)
 	return fmt.Sprintf(`
 %s
 
@@ -522,10 +407,11 @@ resource "azurerm_image" "test" {
     cost-center = "Ops"
   }
 }
-`, r.standaloneImage_setup(data, userName, password, hostName, storageType), storageType == "ZRS", hyperVGenAtt)
+`, template, storageType == "ZRS", hyperVGenAtt)
 }
 
-func (r ImageResource) standaloneImage_requiresImport(data acceptance.TestData, userName string, password string, hostName string) string {
+func (r ImageResource) standaloneImageRequiresImport(data acceptance.TestData) string {
+	template := r.standaloneImageProvision(data, "LRS", "")
 	return fmt.Sprintf(`
 %s
 
@@ -547,214 +433,13 @@ resource "azurerm_image" "import" {
     cost-center = "Ops"
   }
 }
-`, r.standaloneImage_provision(data, userName, password, hostName, "LRS", ""))
+`, template)
 }
 
-func (ImageResource) customImage_fromVHD_setup(data acceptance.TestData, userName string, password string, hostName string) string {
+func (r ImageResource) customImageFromVMWithUnmanagedDisksProvision(data acceptance.TestData) string {
+	template := r.setupUnmanagedDisks(data, "LRS")
 	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-%d"
-  location = "%s"
-}
-
-resource "azurerm_virtual_network" "test" {
-  name                = "acctvn-%d"
-  address_space       = ["10.0.0.0/16"]
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-}
-
-resource "azurerm_subnet" "test" {
-  name                 = "acctsub-%d"
-  resource_group_name  = azurerm_resource_group.test.name
-  virtual_network_name = azurerm_virtual_network.test.name
-  address_prefix       = "10.0.2.0/24"
-}
-
-resource "azurerm_public_ip" "test" {
-  name                = "acctpip-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-  allocation_method   = "Dynamic"
-  domain_name_label   = "%s"
-}
-
-resource "azurerm_network_interface" "testsource" {
-  name                = "acctnicsource-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  ip_configuration {
-    name                          = "testconfigurationsource"
-    subnet_id                     = azurerm_subnet.test.id
-    private_ip_address_allocation = "Dynamic"
-    public_ip_address_id          = azurerm_public_ip.test.id
-  }
-}
-
-resource "azurerm_storage_account" "test" {
-  name                     = "accsa%d"
-  resource_group_name      = azurerm_resource_group.test.name
-  location                 = azurerm_resource_group.test.location
-  account_tier             = "Standard"
-  account_replication_type = "LRS"
-
-  tags = {
-    environment = "Dev"
-  }
-}
-
-resource "azurerm_storage_container" "test" {
-  name                  = "vhds"
-  storage_account_name  = azurerm_storage_account.test.name
-  container_access_type = "blob"
-}
-
-resource "azurerm_virtual_machine" "testsource" {
-  name                  = "testsource"
-  location              = azurerm_resource_group.test.location
-  resource_group_name   = azurerm_resource_group.test.name
-  network_interface_ids = [azurerm_network_interface.testsource.id]
-  vm_size               = "Standard_D1_v2"
-
-  storage_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "18.04-LTS"
-    version   = "latest"
-  }
-
-  storage_os_disk {
-    name          = "myosdisk1"
-    vhd_uri       = "${azurerm_storage_account.test.primary_blob_endpoint}${azurerm_storage_container.test.name}/myosdisk1.vhd"
-    caching       = "ReadWrite"
-    create_option = "FromImage"
-    disk_size_gb  = "30"
-  }
-
-  os_profile {
-    computer_name  = "mdimagetestsource"
-    admin_username = "%s"
-    admin_password = "%s"
-  }
-
-  os_profile_linux_config {
-    disable_password_authentication = false
-  }
-
-  tags = {
-    environment = "Dev"
-    cost-center = "Ops"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, data.RandomInteger, data.RandomInteger, data.RandomInteger, hostName, data.RandomInteger, data.RandomInteger, userName, password)
-}
-
-func (ImageResource) customImage_fromVHD_provision(data acceptance.TestData, userName string, password string, hostName string) string {
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-%d"
-  location = "%s"
-}
-
-resource "azurerm_virtual_network" "test" {
-  name                = "acctvn-%d"
-  address_space       = ["10.0.0.0/16"]
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-}
-
-resource "azurerm_subnet" "test" {
-  name                 = "acctsub-%d"
-  resource_group_name  = azurerm_resource_group.test.name
-  virtual_network_name = azurerm_virtual_network.test.name
-  address_prefix       = "10.0.2.0/24"
-}
-
-resource "azurerm_public_ip" "test" {
-  name                = "acctpip-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-  allocation_method   = "Dynamic"
-  domain_name_label   = "%s"
-}
-
-resource "azurerm_network_interface" "testsource" {
-  name                = "acctnicsource-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  ip_configuration {
-    name                          = "testconfigurationsource"
-    subnet_id                     = azurerm_subnet.test.id
-    private_ip_address_allocation = "Dynamic"
-    public_ip_address_id          = azurerm_public_ip.test.id
-  }
-}
-
-resource "azurerm_storage_account" "test" {
-  name                     = "accsa%d"
-  resource_group_name      = azurerm_resource_group.test.name
-  location                 = azurerm_resource_group.test.location
-  account_tier             = "Standard"
-  account_replication_type = "LRS"
-
-  tags = {
-    environment = "Dev"
-  }
-}
-
-resource "azurerm_storage_container" "test" {
-  name                  = "vhds"
-  storage_account_name  = azurerm_storage_account.test.name
-  container_access_type = "blob"
-}
-
-resource "azurerm_virtual_machine" "testsource" {
-  name                  = "testsource"
-  location              = azurerm_resource_group.test.location
-  resource_group_name   = azurerm_resource_group.test.name
-  network_interface_ids = [azurerm_network_interface.testsource.id]
-  vm_size               = "Standard_D1_v2"
-
-  storage_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "18.04-LTS"
-    version   = "latest"
-  }
-
-  storage_os_disk {
-    name          = "myosdisk1"
-    vhd_uri       = "${azurerm_storage_account.test.primary_blob_endpoint}${azurerm_storage_container.test.name}/myosdisk1.vhd"
-    caching       = "ReadWrite"
-    create_option = "FromImage"
-    disk_size_gb  = "30"
-  }
-
-  os_profile {
-    computer_name  = "mdimagetestsource"
-    admin_username = "%s"
-    admin_password = "%s"
-  }
-
-  os_profile_linux_config {
-    disable_password_authentication = false
-  }
-
-  tags = {
-    environment = "Dev"
-    cost-center = "Ops"
-  }
-}
+%s
 
 resource "azurerm_image" "testdestination" {
   name                = "accteste"
@@ -776,7 +461,7 @@ resource "azurerm_image" "testdestination" {
 }
 
 resource "azurerm_network_interface" "testdestination" {
-  name                = "acctnicdest-%d"
+  name                = "acctnicdest-${local.random}"
   location            = azurerm_resource_group.test.location
   resource_group_name = azurerm_resource_group.test.name
 
@@ -806,8 +491,8 @@ resource "azurerm_virtual_machine" "testdestination" {
 
   os_profile {
     computer_name  = "mdimagetestsource"
-    admin_username = "%s"
-    admin_password = "%s"
+    admin_username = local.admin_username
+    admin_password = local.admin_password
   }
 
   os_profile_linux_config {
@@ -819,44 +504,16 @@ resource "azurerm_virtual_machine" "testdestination" {
     cost-center = "Ops"
   }
 }
-`, data.RandomInteger, data.Locations.Primary, data.RandomInteger, data.RandomInteger, data.RandomInteger, hostName, data.RandomInteger, data.RandomInteger, userName, password, data.RandomInteger, userName, password)
+`, template)
 }
 
-func (ImageResource) customImage_fromVM_sourceVM(data acceptance.TestData, userName string, password string, hostName string) string {
+func (r ImageResource) customImageFromManagedDiskVMProvision(data acceptance.TestData) string {
+	template := r.setupManagedDisks(data)
 	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-%d"
-  location = "%s"
-}
-
-resource "azurerm_virtual_network" "test" {
-  name                = "acctvn-%d"
-  address_space       = ["10.0.0.0/16"]
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-}
-
-resource "azurerm_subnet" "test" {
-  name                 = "acctsub-%d"
-  resource_group_name  = azurerm_resource_group.test.name
-  virtual_network_name = azurerm_virtual_network.test.name
-  address_prefix       = "10.0.2.0/24"
-}
-
-resource "azurerm_public_ip" "test" {
-  name                = "acctpip-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-  allocation_method   = "Dynamic"
-  domain_name_label   = "%s"
-}
+%s
 
 resource "azurerm_network_interface" "testsource" {
-  name                = "acctnicsource-%d"
+  name                = "acctnicsource-${local.number}"
   location            = azurerm_resource_group.test.location
   resource_group_name = azurerm_resource_group.test.name
 
@@ -890,92 +547,8 @@ resource "azurerm_virtual_machine" "testsource" {
 
   os_profile {
     computer_name  = "mdimagetestsource"
-    admin_username = "%s"
-    admin_password = "%s"
-  }
-
-  os_profile_linux_config {
-    disable_password_authentication = false
-  }
-
-  tags = {
-    environment = "Dev"
-    cost-center = "Ops"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, data.RandomInteger, data.RandomInteger, data.RandomInteger, hostName, data.RandomInteger, userName, password)
-}
-
-func (ImageResource) customImage_fromVM_destinationVM(data acceptance.TestData, userName string, password string, hostName string) string {
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-%d"
-  location = "%s"
-}
-
-resource "azurerm_virtual_network" "test" {
-  name                = "acctvn-%d"
-  address_space       = ["10.0.0.0/16"]
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-}
-
-resource "azurerm_subnet" "test" {
-  name                 = "acctsub-%d"
-  resource_group_name  = azurerm_resource_group.test.name
-  virtual_network_name = azurerm_virtual_network.test.name
-  address_prefix       = "10.0.2.0/24"
-}
-
-resource "azurerm_public_ip" "test" {
-  name                = "acctpip-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-  allocation_method   = "Dynamic"
-  domain_name_label   = "%s"
-}
-
-resource "azurerm_network_interface" "testsource" {
-  name                = "acctnicsource-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  ip_configuration {
-    name                          = "testconfigurationsource"
-    subnet_id                     = azurerm_subnet.test.id
-    private_ip_address_allocation = "Dynamic"
-    public_ip_address_id          = azurerm_public_ip.test.id
-  }
-}
-
-resource "azurerm_virtual_machine" "testsource" {
-  name                  = "testsource"
-  location              = azurerm_resource_group.test.location
-  resource_group_name   = azurerm_resource_group.test.name
-  network_interface_ids = [azurerm_network_interface.testsource.id]
-  vm_size               = "Standard_D1_v2"
-
-  storage_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "18.04-LTS"
-    version   = "latest"
-  }
-
-  storage_os_disk {
-    name          = "myosdisk1"
-    caching       = "ReadWrite"
-    create_option = "FromImage"
-  }
-
-  os_profile {
-    computer_name  = "mdimagetestsource"
-    admin_username = "%s"
-    admin_password = "%s"
+    admin_username = local.admin_username
+    admin_password = local.admin_password
   }
 
   os_profile_linux_config {
@@ -989,7 +562,7 @@ resource "azurerm_virtual_machine" "testsource" {
 }
 
 resource "azurerm_image" "testdestination" {
-  name                      = "acctestdest-%d"
+  name                      = "acctestdest-${local.number}"
   location                  = azurerm_resource_group.test.location
   resource_group_name       = azurerm_resource_group.test.name
   source_virtual_machine_id = azurerm_virtual_machine.testsource.id
@@ -1001,7 +574,7 @@ resource "azurerm_image" "testdestination" {
 }
 
 resource "azurerm_network_interface" "testdestination" {
-  name                = "acctnicdest-%d"
+  name                = "acctnicdest-${local.number}"
   location            = azurerm_resource_group.test.location
   resource_group_name = azurerm_resource_group.test.name
 
@@ -1031,8 +604,8 @@ resource "azurerm_virtual_machine" "testdestination" {
 
   os_profile {
     computer_name  = "mdimagetestdest"
-    admin_username = "%s"
-    admin_password = "%s"
+    admin_username = local.admin_username
+    admin_password = local.admin_password
   }
 
   os_profile_linux_config {
@@ -1044,115 +617,11 @@ resource "azurerm_virtual_machine" "testdestination" {
     cost-center = "Ops"
   }
 }
-`, data.RandomInteger, data.Locations.Primary, data.RandomInteger, data.RandomInteger, data.RandomInteger, hostName, data.RandomInteger, userName, password, data.RandomInteger, data.RandomInteger, userName, password)
+`, template)
 }
 
-func (ImageResource) vmss_customImage_fromVHD_setup(data acceptance.TestData, userName string, password string, hostName string) string {
-	return fmt.Sprintf(`
-provider "azurerm" {
-  features {}
-}
-
-resource "azurerm_resource_group" "test" {
-  name     = "acctestRG-%d"
-  location = "%s"
-}
-
-resource "azurerm_virtual_network" "test" {
-  name                = "acctvn-%d"
-  address_space       = ["10.0.0.0/16"]
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-}
-
-resource "azurerm_subnet" "test" {
-  name                 = "acctsub-%d"
-  resource_group_name  = azurerm_resource_group.test.name
-  virtual_network_name = azurerm_virtual_network.test.name
-  address_prefix       = "10.0.2.0/24"
-}
-
-resource "azurerm_public_ip" "test" {
-  name                = "acctpip-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-  allocation_method   = "Dynamic"
-  domain_name_label   = "%s"
-}
-
-resource "azurerm_network_interface" "testsource" {
-  name                = "acctnicsource-%d"
-  location            = azurerm_resource_group.test.location
-  resource_group_name = azurerm_resource_group.test.name
-
-  ip_configuration {
-    name                          = "testconfigurationsource"
-    subnet_id                     = azurerm_subnet.test.id
-    private_ip_address_allocation = "Dynamic"
-    public_ip_address_id          = azurerm_public_ip.test.id
-  }
-}
-
-resource "azurerm_storage_account" "test" {
-  name                     = "accsa%d"
-  resource_group_name      = azurerm_resource_group.test.name
-  location                 = azurerm_resource_group.test.location
-  account_tier             = "Standard"
-  account_replication_type = "LRS"
-
-  tags = {
-    environment = "Dev"
-  }
-}
-
-resource "azurerm_storage_container" "test" {
-  name                  = "vhds"
-  storage_account_name  = azurerm_storage_account.test.name
-  container_access_type = "blob"
-}
-
-resource "azurerm_virtual_machine" "testsource" {
-  name                  = "testsource"
-  location              = azurerm_resource_group.test.location
-  resource_group_name   = azurerm_resource_group.test.name
-  network_interface_ids = [azurerm_network_interface.testsource.id]
-  vm_size               = "Standard_D1_v2"
-
-  storage_image_reference {
-    publisher = "Canonical"
-    offer     = "UbuntuServer"
-    sku       = "18.04-LTS"
-    version   = "latest"
-  }
-
-  storage_os_disk {
-    name          = "myosdisk1"
-    vhd_uri       = "${azurerm_storage_account.test.primary_blob_endpoint}${azurerm_storage_container.test.name}/myosdisk1.vhd"
-    caching       = "ReadWrite"
-    create_option = "FromImage"
-    disk_size_gb  = "30"
-  }
-
-  os_profile {
-    computer_name  = "mdimagetestsource"
-    admin_username = "%s"
-    admin_password = "%s"
-  }
-
-  os_profile_linux_config {
-    disable_password_authentication = false
-  }
-
-  tags = {
-    environment = "Dev"
-    cost-center = "Ops"
-  }
-}
-`, data.RandomInteger, data.Locations.Primary, data.RandomInteger, data.RandomInteger, data.RandomInteger, hostName, data.RandomInteger, data.RandomInteger, userName, password)
-}
-
-func (r ImageResource) vmss_customImage_fromVHD_provision(data acceptance.TestData, userName string, password string, hostName string) string {
-	setup := r.vmss_customImage_fromVHD_setup(data, userName, password, hostName)
+func (r ImageResource) customImageFromVMSSWithUnmanagedDisksProvision(data acceptance.TestData) string {
+	template := r.setupUnmanagedDisks(data, "LRS")
 	return fmt.Sprintf(`
 %s
 
@@ -1188,13 +657,13 @@ resource "azurerm_virtual_machine_scale_set" "testdestination" {
   }
 
   os_profile {
-    computer_name_prefix = "testvm%d"
-    admin_username       = "%s"
-    admin_password       = "%s"
+    computer_name_prefix = "testvm${local.random}"
+    admin_username       = local.admin_username
+    admin_password       = local.admin_password
   }
 
   network_profile {
-    name    = "TestNetworkProfile%d"
+    name    = "TestNetworkProfile"
     primary = true
 
     ip_configuration {
@@ -1214,5 +683,45 @@ resource "azurerm_virtual_machine_scale_set" "testdestination" {
     id = azurerm_image.testdestination.id
   }
 }
-`, setup, data.RandomInteger, userName, password, data.RandomInteger)
+`, template)
+}
+
+func (ImageResource) template(data acceptance.TestData) string {
+	return fmt.Sprintf(`
+locals {
+  number            = "%d"
+  location          = %q
+  domain_name_label = "acctestvm-%s"
+  random_string     = %q
+  admin_username    = "testadmin%d"
+  admin_password    = "Password1234!%d"
+}
+
+resource "azurerm_resource_group" "test" {
+  name     = "acctestRG-${local.number}"
+  location = local.location
+}
+
+resource "azurerm_virtual_network" "test" {
+  name                = "acctvn-${local.number}"
+  resource_group_name = azurerm_resource_group.test.name
+  location            = azurerm_resource_group.test.location
+  address_space       = ["10.0.0.0/16"]
+}
+
+resource "azurerm_subnet" "test" {
+  name                 = "internal"
+  resource_group_name  = azurerm_resource_group.test.name
+  virtual_network_name = azurerm_virtual_network.test.name
+  address_prefix       = "10.0.2.0/24"
+}
+
+resource "azurerm_public_ip" "test" {
+  name                = "acctpip-${local.number}"
+  location            = azurerm_resource_group.test.location
+  resource_group_name = azurerm_resource_group.test.name
+  allocation_method   = "Dynamic"
+  domain_name_label   = local.domain_name_label
+}
+`, data.RandomInteger, data.Locations.Primary, data.RandomString, data.RandomString, data.RandomInteger, data.RandomInteger)
 }
